Add DriverDomain statistics tracking and QEMU test suite
- Implemented `DriverDomainStats` for tracking performance and health metrics of DriverCells, including load duration, uptime, fault counts, and hot swap events.
- Introduced `GlobalDriverDomainStats` for global statistics tracking using atomic operations.
- Created a comprehensive test suite in `tests.rs` to validate the functionality of DriverDomain states, statistics recording, and fault handling.
- Added various smoke tests to ensure the correctness of state transitions, error handling, and statistics updates.
- Included logging and fixture caching for QEMU test exports to facilitate testing in a simulated environment.

+++ kernel/src/net/l4/tcp/mod.rs
@@ -285,11 +285,11 @@ struct TcpSeqState {
     snd_nxt: u32,
     /// 未確認の最古のシーケンス番号
     snd_una: u32,
-    /// 送信ウィンドウサイズ
-    snd_wnd: u16,
+    /// 送信ウィンドウサイズ (scaled)
+    snd_wnd: u32,
     /// 受信シーケンス番号（次に期待するバイト）
     rcv_nxt: u32,
-    /// 受信ウィンドウサイズ
+    /// 受信ウィンドウサイズ (16-bit field for header)
     rcv_wnd: u16,
 }
